feat: 촬영 상태별 카드 분리
